requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
dataclass-factory==2.16
ffmpeg-python==0.2.0
frozenlist==1.7.0
future==1.0.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
multidict==6.6.3
numpy==2.3.2
pillow==11.3.0
propcache==0.3.2
psutil==7.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
shazamio==0.8.1
shazamio_core==1.1.2
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.14.1
yarl==1.20.1
yt-dlp==2025.7.21
